wip: single file spec
