--- packages/myst-spec/src/extensions/footnote.ts
@@ -1,4 +1,0 @@
-export interface FootnoteReferenceExtension {
-  /** @deprecated this should be enumerator */
-  number?: number;
-}
