--- packages/myst-spec/src/nodes/phrasing.ts
@@ -1,11 +1,0 @@
-import type { Parent, PhrasingContent } from 'mdast';
-
-export interface Underline extends Parent {
-  type: 'underline';
-  children: PhrasingContent[];
-}
-
-export interface SmallCaps extends Parent {
-  type: 'smallcaps';
-  children: PhrasingContent[];
-}
